+++ src/main/java/com/codeborne/selenide/commands/GetAttribute.java
@@ -4,8 +4,11 @@ import com.codeborne.selenide.Command;
 import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.impl.WebElementSource;
 
+import javax.annotation.Nullable;
+
 public class GetAttribute implements Command<String> {
   @Override
+  @Nullable
   public String execute(SelenideElement proxy, WebElementSource locator, Object[] args) {
     return locator.getWebElement().getAttribute((String) args[0]);
   }