+++ src/main/java/com/codeborne/selenide/commands/GetInnerText.java
@@ -5,8 +5,11 @@ import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.impl.WebElementSource;
 import org.openqa.selenium.WebElement;
 
+import javax.annotation.Nonnull;
+
 public class GetInnerText implements Command<String> {
   @Override
+  @Nonnull
   public String execute(SelenideElement proxy, WebElementSource locator, Object[] args) {
     WebElement element = locator.getWebElement();
     if (locator.driver().browser().isIE()) {