+++ src/main/java/com/codeborne/selenide/commands/GetText.java
@@ -5,6 +5,8 @@ import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.impl.WebElementSource;
 import org.openqa.selenium.WebElement;
 
+import javax.annotation.Nonnull;
+
 public class GetText implements Command<String> {
   private final GetSelectedText getSelectedText;
 
@@ -17,6 +19,7 @@ public class GetText implements Command<String> {
   }
 
   @Override
+  @Nonnull
   public String execute(SelenideElement proxy, WebElementSource locator, Object[] args) {
     WebElement element = locator.getWebElement();
     return "select".equalsIgnoreCase(element.getTagName()) ?