+++ src/main/java/com/codeborne/selenide/commands/TakeScreenshotAsImage.java
@@ -5,10 +5,12 @@ import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.impl.ScreenShotLaboratory;
 import com.codeborne.selenide.impl.WebElementSource;
 
+import javax.annotation.Nullable;
 import java.awt.image.BufferedImage;
 
 public class TakeScreenshotAsImage implements Command<BufferedImage> {
   @Override
+  @Nullable
   public BufferedImage execute(SelenideElement proxy, WebElementSource element, Object[] args) {
     return ScreenShotLaboratory.getInstance().takeScreenshotAsImage(element.driver(), element.getWebElement());
   }