+++ statics/src/test/java/integration/SelenideMethodsTest.java
@@ -149,10 +149,40 @@ class SelenideMethodsTest extends IntegrationTest {
   }
 
   @Test
-  void userCanSetValueToTextfield() {
+  void innerText_cannotBeNull() {
+    assertThat($("br").innerHtml()).isEqualTo("");
+    assertThatThrownBy(() -> $("#missing").innerHtml())
+      .isInstanceOf(ElementNotFound.class);
+
+    assertThat($("br").innerText()).isEqualTo("");
+    assertThatThrownBy(() -> $("#missing").innerText())
+      .isInstanceOf(ElementNotFound.class);
+  }
+
+  @Test
+  void getValue_canBeEmpty() {
+    assertThat($(By.name("password")).val()).isEqualTo("");
+    assertThat($(By.name("password")).getValue()).isEqualTo("");
+  }
+
+  @Test
+  void getValue_canBeNull() {
+    assertThat($("h2").val()).isNull();
+    assertThat($("h2").getValue()).isNull();
+  }
+
+  @Test
+  void getValue_throwsError_ifElementNotFound() {
+    assertThatThrownBy(() -> $("#missing").val())
+      .isInstanceOf(ElementNotFound.class);
+    assertThatThrownBy(() -> $("#missing").getValue())
+      .isInstanceOf(ElementNotFound.class);
+  }
+
+  @Test
+  void userCanSetValueToTextField() {
     $(By.name("password")).setValue("john");
     $(By.name("password")).val("sherlyn");
-//    $(By.name("password")).shouldBe(focused);
     $(By.name("password")).shouldHave(value("sherlyn"));
     $(By.name("password")).waitUntil(value("sherlyn"), 1000);
     assertThat($(By.name("password")).val())
@@ -562,13 +592,13 @@ class SelenideMethodsTest extends IntegrationTest {
 
   @Test
   void canExecuteJavascript() {
-    long value = (Long) Selenide.executeJavaScript("return 10;");
+    Long value = Selenide.executeJavaScript("return 10;");
     assertThat(value).isEqualTo(10);
   }
 
   @Test
   void canExecuteAsyncJavascript() {
-    long value = Selenide.executeAsyncJavaScript(
+    Long value = Selenide.executeAsyncJavaScript(
       "var callback = arguments[arguments.length - 1]; setTimeout(function() { callback(10); }, 50);"
     );
     assertThat(value).isEqualTo(10);