+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 5.12.2
+* #1172 don't close browser if `holdBrowserOpen=true`  --  see PR #1176
+* #1179 fix @Nonnull annotation for methods getText(), innerText(), innerHtml(), getSelectedText()  --  see PR #1181
+
 ## 5.12.1 (released 25.05.2020)
 * Enable running Selenide without "selenium-ie-driver.jar" "selenium-opera-driver.jar" etc.
 * #1170 fixed Concurrent modification exception in WebDriverFactory  -- see PR #1171